feat: refactor connectorless request
